requirements.txt
slack-sdk==3.34.0
apscheduler==3.10.4
requests==2.32.3
python-dotenv==1.0.1
pytz==2024.2
